requirements.txt
loguru~=0.6.0
lxml~=4.9.2
aiohttp~=3.8.3
PyYAML~=6.0
filetype~=1.2.0